+++ CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Prefer selecting cloud stored videos over disk stored videos.
 
 ### Added
+- [Issue 558](https://github.com/aza547/wow-recorder/issues/488) - Added a bulk delete option.
+
 ### Fixed
 - [Issue 558](https://github.com/aza547/wow-recorder/issues/558) - Fix a bug where it was possible to set your storage path and buffer path to the same thing.
 - [Issue 558](https://github.com/aza547/wow-recorder/issues/558) - Fix a bug where we would rescale the scene too often under some conditions.